LayoutTests/platform/ios-simulator/fast/text/system-font-weight-expected.txt

This test makes sure that extra-heavy weights of the system font can be achieved.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS four.offsetWidth is not three.offsetWidth
FAIL five.offsetWidth should not be 104.
PASS six.offsetWidth is not five.offsetWidth
FAIL eight.offsetWidth should not be 108.
PASS successfullyParsed is true

TEST COMPLETE
Hello, World
Hello, World
Hello, World
Hello, World
Hello, World
Hello, World
Hello, World
Hello, World
Hello, World
